+++ pages/Impedance.py
@@ -16,7 +16,7 @@ from utils import impedance_func as utils_impedance
 ######### Page configuration ######################################################################
 
 st.set_page_config(layout="wide", page_title="SIMsalabim Impedance",
-                   page_icon='./logo/SIMsalabim_logo_HAT.jpg')
+                   page_icon='./Figures/SIMsalabim_logo_HAT.jpg')
 
 # Set the session identifier as query parameter in the URL
 st.query_params.from_dict({'session':st.session_state['id']})
@@ -534,7 +534,7 @@ else:
     # When the reset button is pressed, empty the container and create a List object from the default .txt file. Next, save the default parameters to the parameter file.
     if reset_device_parameters:
         main_container_impedance.empty()
-        dev_par, layers = utils_devpar.load_device_parameters(session_path, zimt_device_parameters, resource_path, True, availLayers=st.session_state['availableLayerFiles'][:-3])
+        dev_par, layers = utils_devpar.load_device_parameters(session_path, zimt_device_parameters, resource_path, True, availLayers=st.session_state['availableLayerFiles'][:-3],run_mode = True)
         save_parameters()
 
     with main_container_impedance.container():
@@ -568,7 +568,9 @@ else:
                     if impedance_item[0] == 'fstep':
                         # Show these parameters as a float
                         impedance_item[1] = st.number_input(impedance_item[0] + '_val', value=impedance_item[1], label_visibility="collapsed")
-                    elif impedance_item[0] == 'V0' or impedance_item[0] == 'delV' or impedance_item[0] == 'G_frac':
+                    elif impedance_item[0] == 'V0':
+                        impedance_item[1] = st.text_input(impedance_item[0] + '_val', value=impedance_item[1], label_visibility="collapsed")
+                    elif impedance_item[0] == 'delV' or impedance_item[0] == 'G_frac':
                         # Show these parameters as a float
                         impedance_item[1] = st.number_input(impedance_item[0] + '_val', value=impedance_item[1], label_visibility="collapsed", format="%f")
                     else:
@@ -686,4 +688,4 @@ else:
     #  Show the SIMsalabim logo in the sidebar
     with st.sidebar:
         st.markdown('<hr>', unsafe_allow_html=True)
-        st.image('./logo/SIMsalabim_logo_cut_trans.png')
+        st.image('./Figures/SIMsalabim_logo_cut_trans.png')