+++ utils/summary_and_citation.py
@@ -88,7 +88,10 @@ def create_summary_and_cite(session_path, used_optics):
             # Impedance
             fp.write(st.session_state['simulation_results'] + ' variables are stored in impedancePars.txt\n')
             fp.write(f'Frequency range: {st.session_state["expObject"]["fmin"]:.2E} to {st.session_state["expObject"]["fmax"]:.2E} Hz\n')
-            fp.write(f'Applied Voltage: {st.session_state["expObject"]["V0"]:.3f} V\n')
+            # if 'oc' in st.session_state["expObject"]["V0"]:
+            fp.write(f'Applied Voltage: {st.session_state["expObject"]["V0"]}\n')
+            # else:
+            #     fp.write(f'Applied Voltage: {st.session_state["expObject"]["V0"]:.3f} V\n')
             fp.write(f'Voltage step size: {st.session_state["expObject"]["delV"]:f} V\n')
             fp.write(f'Fractional Generation rate: {st.session_state["expObject"]["G_frac"]:.3E} (Fraction of the light intensity/generation rate)\n')
 
@@ -97,7 +100,7 @@ def create_summary_and_cite(session_path, used_optics):
             # IMPS
             fp.write(st.session_state['simulation_results'] + ' variables are stored in IMPSPars.txt\n')
             fp.write(f'Frequency range: {st.session_state["expObject"]["fmin"]:.2E} to {st.session_state["expObject"]["fmax"]:.2E} Hz\n')
-            fp.write(f'Applied Voltage: {st.session_state["expObject"]["V0"]:.3f} V\n')
+            fp.write(f'Applied Voltage: {st.session_state["expObject"]["V0"]} V\n')
             fp.write(f'Fractional increase in generation rate: {st.session_state["expObject"]["fracG"]:f} (Fraction to increase the intensity/generation rate with. Sets the size of the initial pertubation)\n')
             fp.write(f'Fractional Generation rate: {st.session_state["expObject"]["G_frac"]:.3E} (Fraction of the light intensity/generation rate)\n')